% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\name{nhanesDXA}
\alias{nhanesDXA}
\title{Import Dual Energy X-ray Absorptiometry (DXA) from 1999-2006}
\usage{
nhanesDXA(year, suppl = FALSE, destfile = NULL)
}
\arguments{
\item{year}{The year of the data to import, where 1999<=year<=2006.}

\item{suppl}{If TRUE then retrieve the supplemental data.}

\item{destfile}{The name of a destination file. If NULL then the data are imported 
into the R environment but no file is created.}
}
\value{
The table is returned as a data frame or else written to file.
}
\description{
Import Dual Energy X-ray Absorptiometry (DXA) from 1999-2006
}
\examples{
\donttest{dxa_b <- nhanesDXA(2001)}
\donttest{dxa_c_s <- nhanesDXA(2003, suppl=TRUE)}
\donttest{nhanesDXA(1999, destfile="dxx.xpt")}
}

