% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_tables.R
\name{nhanesManifest}
\alias{nhanesManifest}
\title{Download and parse the entire NHANES manifest}
\usage{
nhanesManifest(
  which = c("public", "limitedaccess", "variables"),
  sizes = FALSE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{which}{Either "public" or "limitedaccess" to get a manifest
of available tables, or "variables" to get a manifest of
available variables.}

\item{sizes}{Logical, whether to compute data file sizes (as
reported by the server) and include them in the result.}

\item{verbose}{Logical flag indicating whether information on
progress should be reported.}
}
\value{
A data frame, with columns that depend on
  \code{which}. For a manifest of tables, columns are "Table",
  "DocURL", "DataURL", "Years", "Date.Published". If \code{sizes =
  TRUE}, an additional column "DataSize" giving the data file
  sizes in bytes (as reported by the server) is included. For
  limited access tables, the "DataURL" and "DataSize" columns are
  omitted. For a manifest of variables, columns are "VarName",
  "VarDesc", "Table", "TableDesc", "BeginYear",
  "EndYear", "Component", and "UseConstraints".
}
\description{
Downloads and parses the NHANES manifest available at
\url{https://wwwn.cdc.gov/Nchs/Nhanes/search/DataPage.aspx} (for
public data) or
\url{https://wwwn.cdc.gov/Nchs/Nhanes/search/DataPage.aspx?Component=LimitedAccess}
(for limited access data) and returns it as a data frame.
}
\examples{
manifest <- nhanesManifest(sizes = FALSE)
dim(manifest)

}
