% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_codebook.R
\name{nhanesParseCodeBook}
\alias{nhanesParseCodeBook}
\title{Parse NHANES doc URL}
\usage{
nhanesParseCodeBook(url)
}
\arguments{
\item{url}{URL to be downloaded}
}
\value{
list with one element for each variable
}
\description{
Download and parse an NHANES doc file
}
\details{
Downloads and parses an NHANES doc file from a URL and returns it as a list
}
