% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\name{nhanesFromURL}
\alias{nhanesFromURL}
\title{Parse NHANES doc URL}
\usage{
nhanesFromURL(url, translated = TRUE, nchar = 128)
}
\arguments{
\item{url}{URL of XPT file to be downloaded}

\item{translated}{logical, whether variable codes should be translated}

\item{nchar}{integer, labels are truncated after this}
}
\value{
data frame
}
\description{
Download an NHANES table from URL
}
\details{
Downloads an NHANES table from a URL and returns it as a data frame
}
