% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_tables.R
\name{nhanesManifest}
\alias{nhanesManifest}
\title{Returns the entire NHANES manifest}
\usage{
nhanesManifest(which = "Tables")
}
\arguments{
\item{which}{Either "Tables" or "Variables"}
}
\value{
When "Tables" is specified it returns a data.frame with one row for 
each NHANES data table, with columns "Table", "DocURL", "DataURL", "Years", "Date.Published"
}
\description{
Returns the entire NHANES manifest
}
\details{
Will download the comprehensive manifest of variables or tables
}
\examples{
manifest = nhanesManifest()
dim(manifest)

}
