% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nhanes.R
\name{nhanesTables}
\alias{nhanesTables}
\title{Returns a list of table names for the specified survey group.}
\usage{
nhanesTables(nh_surveygroup, year, details = FALSE, namesonly = FALSE,
  includerdconly = FALSE)
}
\arguments{
\item{nh_surveygroup}{The type of survey (DEMOGRAPHIC, DIETARY, EXAMINATION, LABORATORY, QUESTIONNAIRE).
Abbreviated terms may also be used: (DEMO, DIET, EXAM, LAB, Q).}

\item{year}{The year in yyyy format where 1999 <= yyyy <= 2012.}

\item{details}{If TRUE then a more detailed description of the tables is returned.}

\item{namesonly}{If TRUE then only the table names are returned.}

\item{includerdconly}{If TRUE then RDC only tables are included in list.}
}
\value{
The names of the tables in the specified survey group.
}
\description{
Enables quick display of all available tables in the survey group.
}
\details{
Data are retrieved via web scraping using html wrappers from package rvest.
It is often useful to display the table names in an NHANES survey. In effect this
is a convenient way to browse the available NHANES tables.
}
\examples{
nhanesTables('EXAM', 2007)
nhanesTables('LAB', 2009, details=TRUE)
}

